+++ packages/web/app/src/components/layouts/project.tsx
@@ -4,7 +4,6 @@ import { Button } from '@/components/ui/button';
 import { UserMenu } from '@/components/ui/user-menu';
 import { HiveLink } from '@/components/v2/hive-link';
 import { PlusIcon } from '@/components/v2/icon';
-import { CreateTargetModal } from '@/components/v2/modals';
 import { Tabs } from '@/components/v2/tabs';
 import { graphql } from '@/gql';
 import { canAccessProject, ProjectAccessScope, useProjectAccess } from '@/lib/access/project';
@@ -12,6 +11,7 @@ import { useToggle } from '@/lib/hooks';
 import { useLastVisitedOrganizationWriter } from '@/lib/last-visited-org';
 import { Link } from '@tanstack/react-router';
 import { ProjectMigrationToast } from '../project/migration-toast';
+import { CreateTargetModal } from '../ui/modal/create-target';
 import { ProjectSelector } from './project-selector';
 
 export enum Page {
