+++ packages/web/app/src/components/layouts/target.tsx
@@ -2,9 +2,9 @@ import { ReactElement, ReactNode } from 'react';
 import { LinkIcon } from 'lucide-react';
 import { useQuery } from 'urql';
 import { Button } from '@/components/ui/button';
+import { ConnectSchemaModal } from '@/components/ui/modal/connect-schema';
 import { UserMenu } from '@/components/ui/user-menu';
 import { HiveLink } from '@/components/v2/hive-link';
-import { ConnectSchemaModal } from '@/components/v2/modals';
 import { Tabs } from '@/components/v2/tabs';
 import { graphql } from '@/gql';
 import { canAccessTarget, TargetAccessScope, useTargetAccess } from '@/lib/access/target';
