+++ packages/web/app/src/components/ui/modal/connect-schema.tsx
@@ -1,0 +1,230 @@
+import { ReactElement, useMemo, useState } from 'react';
+import { useQuery } from 'urql';
+import { Button } from '@/components/ui/button';
+import {
+  Dialog,
+  DialogContent,
+  DialogFooter,
+  DialogHeader,
+  DialogTitle,
+} from '@/components/ui/dialog';
+import { Heading } from '@/components/ui/heading';
+import { CopyValue, DocsLink, Link, RadixSelect, Tag } from '@/components/v2';
+import { graphql } from '@/gql';
+import { ProjectType } from '@/gql/graphql';
+import { getDocsUrl } from '@/lib/docs-url';
+
+const ConnectSchemaModalQuery = graphql(`
+  query ConnectSchemaModal($targetSelector: TargetSelectorInput!) {
+    target(selector: $targetSelector) {
+      id
+      project {
+        id
+        type
+      }
+      cdnUrl
+      activeContracts(first: 20) {
+        edges {
+          node {
+            id
+            contractName
+            cdnUrl
+          }
+        }
+      }
+    }
+  }
+`);
+
+type CdnArtifactType = 'sdl' | 'services' | 'supergraph' | 'metadata';
+
+const ArtifactToProjectTypeMapping: Record<ProjectType, CdnArtifactType[]> = {
+  [ProjectType.Single]: ['sdl', 'metadata'],
+  [ProjectType.Stitching]: ['sdl', 'services'],
+  [ProjectType.Federation]: ['sdl', 'services', 'supergraph'],
+};
+
+const ArtifactTypeToDisplayName: Record<CdnArtifactType, string> = {
+  sdl: 'GraphQL SDL',
+  services: 'Services Definition and SDL',
+  supergraph: 'Apollo Federation Supergraph',
+  metadata: 'Hive Schema Metadata',
+};
+
+function composeEndpoint(baseUrl: string, artifactType: CdnArtifactType): string {
+  return `${baseUrl}/${artifactType}`;
+}
+
+export const ConnectSchemaModal = (props: {
+  isOpen: boolean;
+  toggleModalOpen: () => void;
+  organizationId: string;
+  projectId: string;
+  targetId: string;
+}): ReactElement => {
+  const { isOpen, toggleModalOpen } = props;
+  const [selectedGraph, setSelectedGraph] = useState<string>('DEFAULT_GRAPH');
+  const [selectedArtifact, setSelectedArtifact] = useState<CdnArtifactType>('sdl');
+  const [query] = useQuery({
+    query: ConnectSchemaModalQuery,
+    variables: {
+      targetSelector: {
+        organization: props.organizationId,
+        project: props.projectId,
+        target: props.targetId,
+      },
+    },
+    requestPolicy: 'cache-and-network',
+    // we only need to fetch the data when the modal is open
+    pause: !isOpen,
+  });
+
+  const selectedContract = useMemo(() => {
+    if (selectedGraph === 'DEFAULT_GRAPH') {
+      return null;
+    }
+    return query.data?.target?.activeContracts.edges.find(
+      ({ node }) => node.contractName === selectedGraph,
+    )?.node;
+  }, [selectedGraph, query.data?.target?.activeContracts.edges]);
+
+  return (
+    <Dialog open={isOpen} onOpenChange={toggleModalOpen}>
+      <DialogContent className="flex w-[800px] flex-col gap-5">
+        <DialogHeader>
+          <DialogTitle>
+            <Heading className="text-center">Hive CDN Access</Heading>
+          </DialogTitle>
+        </DialogHeader>
+        {query.data?.target && (
+          <>
+            <p className="text-sm text-gray-500">
+              Hive leverages the{' '}
+              <Link
+                as="a"
+                variant="primary"
+                className="font-bold underline"
+                href="https://www.cloudflare.com/network"
+                target="_blank"
+                rel="noreferrer"
+              >
+                CloudFlare Global Network
+              </Link>{' '}
+              to deliver your GraphQL schema and schema metadata. This means that your schema will
+              be available from the nearest location to your GraphQL gateway, with 100% uptime,
+              regardless of Hive's status.
+            </p>
+            <p className="text-sm text-gray-500">
+              Based on your project type, you can access different artifacts from Hive's CDN:
+            </p>
+            <div>
+              <RadixSelect
+                className="mr-2"
+                placeholder="Select Graph"
+                name="graph-variant-select"
+                position="popper"
+                value={selectedGraph}
+                options={[
+                  {
+                    value: 'DEFAULT_GRAPH',
+                    label: 'Default Graph',
+                  },
+                  ...query.data.target.activeContracts.edges.map(({ node }) => ({
+                    value: node.contractName,
+                    label: node.contractName,
+                  })),
+                ]}
+                onChange={value => {
+                  if (
+                    value !== 'DEFAULT_GRAPH' &&
+                    selectedArtifact !== 'sdl' &&
+                    selectedArtifact !== 'supergraph'
+                  ) {
+                    setSelectedArtifact('sdl');
+                  }
+                  setSelectedGraph(value);
+                }}
+              />
+              <RadixSelect
+                placeholder="Select Artifact"
+                name="artifact-select"
+                position="popper"
+                value={selectedArtifact}
+                options={ArtifactToProjectTypeMapping[query.data.target.project.type].map(t => ({
+                  value: t,
+                  label: ArtifactTypeToDisplayName[t],
+                  disabled: t !== 'supergraph' && t !== 'sdl' && selectedGraph !== 'DEFAULT_GRAPH',
+                }))}
+                onChange={setSelectedArtifact}
+              />
+            </div>
+            <span className="text-sm text-gray-500">
+              To access your schema from Hive's CDN, use the following endpoint:
+            </span>
+            <CopyValue
+              value={composeEndpoint(
+                selectedContract?.cdnUrl ?? query.data.target.cdnUrl,
+                selectedArtifact,
+              )}
+            />
+            <span className="text-sm text-gray-500">
+              To authenticate,{' '}
+              <Link
+                search={{
+                  page: 'cdn',
+                }}
+                variant="primary"
+                className="font-bold underline"
+                to="/$organizationId/$projectId/$targetId/settings"
+                params={{
+                  organizationId: props.organizationId,
+                  projectId: props.projectId,
+                  targetId: props.targetId,
+                }}
+                target="_blank"
+                rel="noreferrer"
+              >
+                create a CDN Access Token from your target's Settings page
+              </Link>{' '}
+              use the CDN access token in your HTTP headers:
+              <br />
+            </span>
+            <p className="text-sm text-gray-500">
+              <Tag>
+                X-Hive-CDN-Key: {'<'}Your Access Token{'>'}
+              </Tag>
+            </p>
+            <DocsLink href="/features/high-availability-cdn">
+              Learn more about Hive High-Availability CDN
+            </DocsLink>
+            {query.data.target.project.type === ProjectType.Federation ? (
+              <p className="text-sm text-gray-500">
+                Read the{' '}
+                <Link
+                  variant="primary"
+                  target="_blank"
+                  rel="noreferrer"
+                  to={getDocsUrl('/integrations/apollo-gateway#supergraph-sdl-from-the-cdn')}
+                >
+                  Using the Registry with a Apollo Gateway
+                </Link>{' '}
+                chapter in our documentation.
+              </p>
+            ) : null}
+          </>
+        )}
+        <DialogFooter>
+          <Button
+            type="button"
+            variant="secondary"
+            size="lg"
+            onClick={toggleModalOpen}
+            className="self-end"
+          >
+            Close
+          </Button>
+        </DialogFooter>
+      </DialogContent>
+    </Dialog>
+  );
+};
