+++ packages/web/app/src/pages/organization-settings.tsx
@@ -17,14 +17,12 @@ import { DocsLink } from '@/components/ui/docs-note';
 import { Form, FormControl, FormField, FormItem, FormMessage } from '@/components/ui/form';
 import { Input } from '@/components/ui/input';
 import { Meta } from '@/components/ui/meta';
+import { DeleteOrganizationModal } from '@/components/ui/modal/delete-organization';
 import { Subtitle, Title } from '@/components/ui/page';
 import { QueryError } from '@/components/ui/query-error';
 import { useToast } from '@/components/ui/use-toast';
 import { GitHubIcon, SlackIcon } from '@/components/v2/icon';
-import {
-  DeleteOrganizationModal,
-  TransferOrganizationOwnershipModal,
-} from '@/components/v2/modals';
+import { TransferOrganizationOwnershipModal } from '@/components/v2/modals';
 import { Tag } from '@/components/v2/tag';
 import { env } from '@/env/frontend';
 import { FragmentType, graphql, useFragment } from '@/gql';
