+++ packages/web/app/src/pages/target-settings.tsx
@@ -13,6 +13,8 @@ import { CardDescription } from '@/components/ui/card';
 import { Checkbox } from '@/components/ui/checkbox';
 import { DocsLink } from '@/components/ui/docs-note';
 import { Meta } from '@/components/ui/meta';
+import { CreateAccessTokenModal } from '@/components/ui/modal/create-access-token';
+import { DeleteTargetModal } from '@/components/ui/modal/delete-target';
 import {
   NavLayout,
   PageLayout,
@@ -25,7 +27,6 @@ import { TimeAgo } from '@/components/ui/time-ago';
 import { useToast } from '@/components/ui/use-toast';
 import { Combobox } from '@/components/v2/combobox';
 import { Input } from '@/components/v2/input';
-import { CreateAccessTokenModal, DeleteTargetModal } from '@/components/v2/modals';
 import { Spinner } from '@/components/v2/spinner';
 import { Switch } from '@/components/v2/switch';
 import { Table, TBody, Td, Tr } from '@/components/v2/table';
